+++ CHANGELOG.md
@@ -25,6 +25,9 @@ All notable changes to this project will be documented in this file.
   - multiple instructions found when only one allowed
     - CMD ([#81](https://github.com/rcjsuen/dockerfile-language-server-nodejs/issues/81))
     - HEALTHCHECK ([#80](https://github.com/rcjsuen/dockerfile-language-server-nodejs/issues/80))
+  - check that the same flag is not used twice ([#83](https://github.com/rcjsuen/dockerfile-language-server-nodejs/issues/83))
+    - COPY's from
+    - HEALTHCHECK's interval, retries, start-period, timeout
   - HEALTHCHECK
     - warn if arguments follow a HEALTHCHECK NONE ([#84](https://github.com/rcjsuen/dockerfile-language-server-nodejs/issues/84))
     - warn if the retries flag doesn't specify a number ([#85](https://github.com/rcjsuen/dockerfile-language-server-nodejs/issues/85))